使用cstdint定义的数据类型重命名了相关的变量类型，此外还修正了命名规则，现在全都使用驼峰命名规则，放弃了c语言的类型转换方法现在一律都使用***_cast来转换类型

+++ Engine/Inc/Gear/Core/Resource/BufferView.h
@@ -18,7 +18,7 @@ class BufferView :public EngineResource
 {
 public:
 
-	BufferView(Buffer* const buffer, const UINT structureByteStride, const DXGI_FORMAT format, const UINT size, const bool createSRV, const bool createUAV, const bool createVBV, const bool createIBV, const bool cpuWritable, const bool persistent);
+	BufferView(Buffer* const buffer, const uint32_t structureByteStride, const DXGI_FORMAT format, const uint64_t size, const bool createSRV, const bool createUAV, const bool createVBV, const bool createIBV, const bool cpuWritable, const bool persistent);
 
 	~BufferView();
 
@@ -48,9 +48,9 @@ private:
 
 	CounterBufferView* counterBuffer;
 
-	UINT srvIndex;
+	uint32_t srvIndex;
 
-	UINT uavIndex;
+	uint32_t uavIndex;
 
 	D3D12_GPU_DESCRIPTOR_HANDLE viewGPUHandle;
 
@@ -65,7 +65,7 @@ private:
 
 	UploadHeap** uploadHeaps;
 
-	UINT uploadHeapIndex;
+	uint32_t uploadHeapIndex;
 
 	Buffer* buffer;
 };
